Various changes
Added manifesto and other minor content.

Modified `config.yaml` to enable features and setup the basic website
structure.

Added logo and minor stylistic changes to the basic theme.

Note that the file `layouts/_default/list.html` is copied directly from
the theme, with a very minor change to allow more pages to appear on the
homepage. There is probably a neater way of doing this, such as forking
the theme.